Move fragment discovery to a dedicated module
