+++ tests/units_test.py
@@ -4,14 +4,10 @@ import sys
 import pytest
 
 from sphinxcontrib.towncrier._compat import shlex_join
-from sphinxcontrib.towncrier.ext import (
-    _find_config_file, _get_changelog_draft_entries,
-)
+from sphinxcontrib.towncrier.ext import _get_changelog_draft_entries
 
 
 NO_OUTPUT_MARKER = r'\[No output\]'
-PYPROJECT_TOML_FILENAME = 'pyproject.toml'
-TOWNCRIER_TOML_FILENAME = 'towncrier.toml'
 
 
 _get_changelog_draft_entries_unwrapped = (
@@ -100,42 +96,3 @@ def test_towncrier_draft_generation_failure_msg(
 
     with pytest.raises(RuntimeError, match=expected_error_message):
         _get_changelog_draft_entries_unwrapped(version_string)
-
-
-@pytest.mark.parametrize(
-    ('config_file_names_on_disk', 'expected_config_file_name'),
-    (
-        pytest.param(
-            set(),
-            PYPROJECT_TOML_FILENAME,
-            id='pyproject.toml-when-no-configs',
-        ),
-        pytest.param(
-            {PYPROJECT_TOML_FILENAME},
-            'pyproject.toml',
-            id='pyproject.toml-only',
-        ),
-        pytest.param(
-            {TOWNCRIER_TOML_FILENAME},
-            TOWNCRIER_TOML_FILENAME,
-            id='towncrier.toml-only',
-        ),
-        pytest.param(
-            {PYPROJECT_TOML_FILENAME, TOWNCRIER_TOML_FILENAME},
-            TOWNCRIER_TOML_FILENAME,
-            id='towncrier.toml-over-pyproject.toml',
-        ),
-    ),
-)
-def test_find_config_files(
-        config_file_names_on_disk,
-        expected_config_file_name,
-        tmp_path,
-) -> None:
-    """Verify that the correct Towncrier config is always preferred."""
-    for config_file_name_on_disk in config_file_names_on_disk:
-        tmp_path.joinpath(config_file_name_on_disk).write_text(
-            '', encoding='utf-8',
-        )
-
-    assert _find_config_file(tmp_path).name == expected_config_file_name
